refactor: update import paths and remove fullscreen flag
