+++ src/components/message-input/styles.scss
@@ -102,14 +102,10 @@
 
   &__emoji-picker-container {
     position: absolute;
-    right: 74px;
+    right: 65px;
     bottom: 75px;
     z-index: 12;
     color: theme.$color-greyscale-11;
-
-    &--fullscreen {
-      right: 65px;
-    }
   }
 
   &__disabled-tooltip {
